3 iteration

+++ ConsoleApp/SmoLibrary/Static.cs
@@ -1,4 +1,5 @@
-﻿using System.IO.Compression;
+﻿using System.Globalization;
+using System.IO.Compression;
 
 //Der Iris-Datensatz (nur zwei Klassen, z. B. Setosa vs. Versicolor) ist linear separierbar bei den Features sepal length und petal length.
 namespace ClassLibrary1;
@@ -14,42 +15,52 @@ public static class Static
     {
         return u.Select(x => x.InnerProduct(v));
     }
-    
-    
+
+
     public static double DoLogic(string fileName, double split)
     {
         IEnumerable<string> lines = File.ReadLines(fileName); // Lazily read lines
-        var result = lines.Skip(1)
+        (double[] filteredInput, string label)[] result = lines.Skip(1)
             .Select(line =>
             {
                 string[] r = line.Split(',');
                 //Iris-setosa
                 //Iris-versicolor
                 //Iris-virginica
-                var input = r[1..^1].Select(Convert.ToDouble).ToArray();
-                var filteredInput = new double[] { input[0], input[2] };
+                double[] input = r[1..^1].Select(Convert.ToDouble).ToArray();
+                double[] filteredInput = new[] { input[0], input[2] };
                 string label = r[^1];
                 return (filteredInput, label);
             }).ToArray();
 
         Random r = new();
-        result = result.Where(x => x.label is "Iris-setosa" or "Iris-versicolo").ToArray();
-        r.Shuffle(result);
+        var sampleBeing = 45;
+        var sampleSize = 10;
+        
+        result = result.Where(x => x.label is "Iris-setosa" or "Iris-versicolor").Skip(sampleBeing).Take(10).ToArray();
+        //r.Shuffle(result);
 
         int length = (int)(result.Length * split);
-        var train = result.Take(length);
-        var test = result.Skip(length).ToList();
+        IEnumerable<(double[] filteredInput, string label)> train = result.Take(length);
+        List<(double[] filteredInput, string label)> test = result.Skip(length).ToList();
         OneVsAllClassifier oneVsAllClassifier = new(train);
 
         oneVsAllClassifier.fit();
         int correctCount = test.Count(x => oneVsAllClassifier.Predict(x.filteredInput) == x.label);
-        var finalR = (double)correctCount / test.Count;
+        double finalR = (double)correctCount / test.Count;
         File.WriteAllText($"result_{r.Next()}_{DateTime.Now.Minute}", $"Result %: {finalR}");
         Console.WriteLine(finalR);
+        foreach (KeyValuePair<string, SvmOptimizer> VARIABLE in oneVsAllClassifier.Smos)
+        {
+            Console.WriteLine(VARIABLE.Key);
+            Console.WriteLine(string.Join(" ", VARIABLE.Value.W.Select(x => x.ToString(CultureInfo.InvariantCulture))));
+            Console.WriteLine(VARIABLE.Value.B);
+            Console.WriteLine();
+        }
+
         return finalR;
     }
-    
-    
+
 
     public static async Task<string> KaggleDownload(string username, string dataName)
     {
@@ -68,10 +79,7 @@ public static class Static
 
         await using FileStream fileStreamRead = File.OpenRead(zipName);
 
-        if (Directory.Exists(dataName))
-        {
-            Directory.Delete(dataName, true);
-        }
+        if (Directory.Exists(dataName)) Directory.Delete(dataName, true);
         Directory.CreateDirectory(dataName);
 
         ZipFile.ExtractToDirectory(fileStreamRead, dataName);
